fakesensor/requirements.txt
requests
google-cloud-pubsub
google-cloud-datastore
